Make better use of Scala language features.
This is a big commit that rewrites most of the original code.
The original code was the first Scala code I wrote. This commit updates this
first version with lots a new language features I learned to use in the meanwhile:
case classes, pattern matching, collection api features, usage of companion object,...

Another change is also that I use immutable data structures which had mainly an impact
on the Grid implementation (formerly called Board).